Refactor: auth client component 생성
