+++ src/app/page.tsx
@@ -1,26 +1,5 @@
-"use client";
-
-import Image from "next/image";
-import login from "/public/images/login.svg";
-import LoginForm from "@/components/LoginForm/LoginForm";
-import { auth } from "@/utils/firebase";
-import { useEffect } from "react";
-import { useRouter } from "next/navigation";
+import Auth from "@/components/Auth/Auth";
 
 export default function LoginPage() {
-  const router = useRouter();
-  useEffect(() => {
-    auth.onAuthStateChanged(user => {
-      if (user) {
-        router.push("/community");
-      }
-    });
-  }, []);
-
-  return (
-    <div className="h-screen flex flex-col justify-center items-center gap-y-12">
-      <Image src={login} alt="logo" priority={true} />
-      <LoginForm />
-    </div>
-  );
+  return <Auth />;
 }
